requirements.txt.txt
requests
pandas
beautifulsoup4
psycopg2
pymongo
python-dotenv
